+++ module/applications/apps/presence-attack.mjs
@@ -200,7 +200,7 @@ export class PresenceAttackApplication extends HeroApplication {
 
         const cardData = {
             tags: tags.map((tag) => {
-                return { ...tag, value: tag.value.signedStringHero() };
+                return { ...tag, value: tag.value.signedString() };
             }),
             rolls: preAttackRoller.rawRolls(),
             renderedRoll: rollHtml,