+++ module/item/skill.mjs
@@ -307,7 +307,7 @@ async function skillRoll(item, actor, target) {
     // render card
     const cardData = {
         tags: tags.map((tag) => {
-            return { ...tag, value: tag.value.signedStringHero() };
+            return { ...tag, value: tag.value.signedString() };
         }),
         rolls: skillRoller.rawRolls(),
         renderedRoll: rollHtml,